# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Analysis Canvas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/load-cases/{load_case_id}/drop-videos": {
      "get": {
        "summary": "Get Drop Videos",
        "operationId": "get_drop_videos_api_load_cases__load_case_id__drop_videos_get",
        "parameters": [
          {
            "name": "load_case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Load Case Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DropVideoPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboards/{dashboard_id}": {
      "get": {
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_api_dashboards__dashboard_id__get",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Api Dashboards  Dashboard Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Dashboard",
        "operationId": "save_dashboard_api_dashboards__dashboard_id__put",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Dashboard Api Dashboards  Dashboard Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboards/{dashboard_id}/versions/{version}": {
      "get": {
        "summary": "Get Dashboard Version",
        "operationId": "get_dashboard_version_api_dashboards__dashboard_id__versions__version__get",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "include_invalid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Invalid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Version Api Dashboards  Dashboard Id  Versions  Version  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Dashboard Version",
        "operationId": "delete_dashboard_version_api_dashboards__dashboard_id__versions__version__delete",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Dashboard Version Api Dashboards  Dashboard Id  Versions  Version  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{dashboard_id}/clone": {
      "post": {
        "summary": "Clone Dashboard",
        "operationId": "clone_dashboard_api_dashboards__dashboard_id__clone_post",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardClone"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clone Dashboard Api Dashboards  Dashboard Id  Clone Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{dashboard_id}/restore/{version}": {
      "post": {
        "summary": "Restore Dashboard",
        "operationId": "restore_dashboard_api_dashboards__dashboard_id__restore__version__post",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Restore Dashboard Api Dashboards  Dashboard Id  Restore  Version  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisPageMeta": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "analysis_page",
            "title": "Kind",
            "default": "analysis_page"
          },
          "analysis_key": {
            "type": "string",
            "enum": [
              "open_cell",
              "chassis_rear",
              "run_comparison",
              "custom"
            ],
            "title": "Analysis Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "title": "Status"
          },
          "display_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Display Order"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          }
        },
        "type": "object",
        "required": [
          "analysis_key",
          "status",
          "display_order",
          "is_system"
        ],
        "title": "AnalysisPageMeta"
      },
...
      "DashboardClone": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DashboardClone"
      },
      "DashboardDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/Widget"
            },
            "type": "array",
            "title": "Widgets"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisPageMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "widgets"
        ],
        "title": "DashboardDefinition"
      },
...
      "DropVideoEvaluation": {
        "properties": {
          "overall_verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL"
            ],
            "title": "Overall Verdict"
          },
          "open_cell": {
            "$ref": "#/components/schemas/DropVideoSubsystemEvaluation"
          },
          "chassis_rear": {
            "$ref": "#/components/schemas/DropVideoSubsystemEvaluation"
          }
        },
        "type": "object",
        "required": [
          "overall_verdict",
          "open_cell",
          "chassis_rear"
        ],
        "title": "DropVideoEvaluation"
      },
      "DropVideoEvaluationSummary": {
        "properties": {
          "total_scenes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Scenes"
          },
          "pass_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pass Count"
          },
          "fail_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Fail Count"
          },
          "open_cell": {
            "$ref": "#/components/schemas/DropVideoSubsystemSummary"
          },
          "chassis_rear": {
            "$ref": "#/components/schemas/DropVideoSubsystemSummary"
          }
        },
        "type": "object",
        "required": [
          "total_scenes",
          "pass_count",
          "fail_count",
          "open_cell",
          "chassis_rear"
        ],
        "title": "DropVideoEvaluationSummary"
      },
      "DropVideoItemResponse": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "scene_name": {
            "type": "string",
            "title": "Scene Name"
          },
          "video_url": {
            "type": "string",
            "title": "Video Url"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "file_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "File Size"
          },
          "format": {
            "type": "string",
            "enum": [
              "mp4",
              "webm"
            ],
            "title": "Format"
          },
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "fast_start": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fast Start"
          },
          "sort_order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sort Order"
          },
          "drop_direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drop Direction"
          },
          "drop_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drop Condition"
          },
          "analysis_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Version"
          },
          "evaluation": {
            "$ref": "#/components/schemas/DropVideoEvaluation"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "scene_id",
          "scene_name",
          "video_url",
          "thumbnail_url",
          "file_size",
          "format",
          "codec",
          "fast_start",
          "sort_order",
          "drop_direction",
          "drop_condition",
          "analysis_version",
          "evaluation"
        ],
        "title": "DropVideoItemResponse"
      },
      "DropVideoLoadCaseResponse": {
        "properties": {
          "load_case_id": {
            "type": "string",
            "title": "Load Case Id"
          },
          "load_case_name": {
            "type": "string",
            "title": "Load Case Name"
          },
          "analysis_type": {
            "type": "string",
            "title": "Analysis Type"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "request_name": {
            "type": "string",
            "title": "Request Name"
          }
        },
        "type": "object",
        "required": [
          "load_case_id",
          "load_case_name",
          "analysis_type",
          "request_id",
          "request_name"
        ],
        "title": "DropVideoLoadCaseResponse"
      },
      "DropVideoPageResponse": {
        "properties": {
          "load_case": {
            "$ref": "#/components/schemas/DropVideoLoadCaseResponse"
          },
          "source": {
            "type": "string",
            "enum": [
              "DATABASE",
              "EXAMPLE_ADAPTER"
            ],
            "title": "Source"
          },
          "demo_only": {
            "type": "boolean",
            "title": "Demo Only"
          },
          "evaluation_source": {
            "type": "string",
            "const": "SYNTHETIC_DEMO",
            "title": "Evaluation Source"
          },
          "contract_version": {
            "type": "integer",
            "const": 1,
            "title": "Contract Version"
          },
          "summary": {
            "$ref": "#/components/schemas/DropVideoEvaluationSummary"
          },
          "pagination": {
            "$ref": "#/components/schemas/DropVideoPaginationResponse"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/DropVideoItemResponse"
            },
            "type": "array",
            "title": "Videos"
          }
        },
        "type": "object",
        "required": [
          "load_case",
          "source",
          "demo_only",
          "evaluation_source",
          "contract_version",
          "summary",
          "pagination",
          "videos"
        ],
        "title": "DropVideoPageResponse"
      },
      "DropVideoPaginationResponse": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_previous",
          "has_next"
        ],
        "title": "DropVideoPaginationResponse"
      },
      "DropVideoSubsystemEvaluation": {
        "properties": {
          "critical_value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Critical Value"
          },
          "threshold": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Threshold"
          },
          "unit": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Unit"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL"
            ],
            "title": "Verdict"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "maxProperties": 8,
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "critical_value",
          "threshold",
          "unit",
          "verdict"
        ],
        "title": "DropVideoSubsystemEvaluation"
      },
      "DropVideoSubsystemSummary": {
        "properties": {
          "pass_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pass Count"
          },
          "fail_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Fail Count"
          },
          "threshold": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Threshold"
          },
          "unit": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "pass_count",
          "fail_count",
          "threshold",
          "unit"
        ],
        "title": "DropVideoSubsystemSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Widget": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "kpi",
              "verdict",
              "gauge",
              "edge_bar",
              "time_series",
              "scatter",
              "result_table",
              "contour",
              "video",
              "video_grid",
              "model3d",
              "note",
              "workflow",
              "open_cell_map",
              "open_cell_summary",
              "summary",
              "chassis_summary",
              "chassis_diagram",
              "chassis_bar",
              "chassis_table",
              "run_comparison"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "w": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "W"
          },
          "h": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "H"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "Widget"
      },
...
    }
  }
}